moved device handling in own file
